Add JSON tags.
This ensures the names are in the correct case, and, were
Control/Player/Playlist set to nil, their absence.  Currently, this is
never the case, but in future when we check features it will be.

Incidentally, unsure how to make Time/File go away only when their
respective features are not available.  Maybe we need to make some sort
of Maybe/Option-esque wrapper for them.