Use maps for dispatching resource requests.
This gets rid of the resource structs entirely.  Instead, for now, each
(basic composite) resource is described entirely by a map from the names
of its children to another handling function that works on requests in
that resource.

Sounds utterly overcomplicated, but it basically just means that if you
ask for "foo/bar/baz", it'll go through a map {"foo": fooGet, ...} and
fooGet will be called with "bar/baz", and so on.  bazGet would then
actually either be some resource, or it'll have its own map, see that
you didn't ask for anything in the map, and just send you the whole
thing.

It certainly saves on code, should be extensible to things that aren't
GETs, and works fairly well.